+++ CHANGELOG.md
@@ -23,6 +23,21 @@ If you like this project and find it useful, please consider giving it a star on
   <img src="bmc-button.svg" alt="Buy me a coffee" width="120">
 </a>
 
+## [1.3.1] - 2025-06-??
+
+### Added
+
+- [platform]: Added solarPower device type (not supported by the Home app).
+- [platform]: Added batteryStoraga device type (not supported by the Home app).
+
+### Changed
+
+- [platform]: Changed imports from matterbridge/devices.
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="bmc-button.svg" alt="Buy me a coffee" width="80">
+</a>
+
 ## [1.3.0] - 2025-06-25
 
 ### Added